# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de gestion des catégories",
        "description": "Documentation pour l'API de gestion des catégories de votre application.",
        "contact": {
            "email": "votre_email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tickets/assign": {
            "post": {
                "tags": [
                    "Attribution des Tickets"
                ],
                "summary": "Attribue un ensemble de tickets à un utilisateur.",
                "description": "Cette méthode lit l'ID de l'utilisateur et une liste d'IDs de tickets pour créer des entrées dans la table 'ticketagent'.",
                "operationId": "9a5558abb414be5a4b9a48d1b347e8c3",
                "requestBody": {
                    "description": "Données d'attribution (user_id et ticketinstance_ids)",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "ticketinstance_ids"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID de l'utilisateur à qui attribuer les tickets.",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "ticketinstance_ids": {
                                        "description": "Liste des IDs des tickets à attribuer.",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 101
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tickets attribués avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tickets attribués avec succès à l'utilisateur."
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "assigned_tickets": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer",
                                                "example": 101
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé."
                    },
                    "422": {
                        "description": "Erreur de validation des données (champs manquants ou invalides)."
                    }
                }
            }
        },
        "/api/users/assigned-tickets": {
            "get": {
                "tags": [
                    "Attribution des Tickets"
                ],
                "summary": "Lister tous les utilisateurs et leurs tickets attribués.",
                "description": "Récupère une liste de tous les utilisateurs qui ont au moins un ticket attribué, avec les détails de ces tickets.",
                "operationId": "161b356bdeba8ce4ddf7c6788e0b8d15",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs et de leurs tickets.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun utilisateur n'a de ticket attribué."
                    }
                }
            }
        },
        "/api/me/assigned-tickets": {
            "get": {
                "tags": [
                    "Attribution des Tickets"
                ],
                "summary": "Récupérer les tickets attribués à l'utilisateur connecté.",
                "description": "Récupère la liste des Ticketinstance attribuées à l'utilisateur qui fait l'appel (authentifié).",
                "operationId": "b7f221967f29bbbd0276007dd3596853",
                "responses": {
                    "200": {
                        "description": "Liste des tickets attribués à l'utilisateur.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié."
                    },
                    "404": {
                        "description": "Aucun ticket attribué à cet utilisateur."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}